feat: added main menu and pause menu
Made a new node as the main scene and assigned the main menu, pause menu, and work scene into it. Game won't start before the start button is pressed. 
Game will be paused when the top right button is pressed and will continue when the resume button is pressed. Game will end if the quit button is pressed.