feat: Added main menu and pause menu
Made a new node to set as the main scene and put the work, start, and pause scene in it. Game won't start until start button is pressed. Game will pause when the top right button is pressed and will continue when the resume button is pressed. Game will end when the quit button is pressed whether from start or pause menu.